+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-04-16 15:05:55
-// Commit: perf(ui): apply RepaintBoundary on animated ui widgets
+// Updated at: 2026-04-16 16:54:06
+// Commit: refactor(ui): extract reusable atomic widgets from ui
 
 class uiComponentState {
-  final String commitRef = '425';
-  final String lastMessage = 'perf(ui): apply RepaintBoundary on animated ui widgets';
-  final DateTime timestamp = DateTime.parse('2026-04-16T15:05:55+05:00');
+  final String commitRef = '432';
+  final String lastMessage = 'refactor(ui): extract reusable atomic widgets from ui';
+  final DateTime timestamp = DateTime.parse('2026-04-16T16:54:06+05:00');
 }